@@ -8,9 +8,9 @@ class PingController < ApplicationController
 
   def verify
     if ActiveRecord::Base.connected?
-      render text: 'Application is OK'
+      render plain: 'Application is OK'
     else
-      render text: 'Cannot connect to database!', status: :service_unavailable
+      render plain: 'Cannot connect to database!', status: :service_unavailable
     end
   end
 end
