@@ -1,4 +1,4 @@
-class ConfigLinks < ActiveRecord::Migration
+class ConfigLinks < ActiveRecord::Migration[4.2]
   def change
 		create_table "links", force: :cascade do |t|
 			t.string   "url"
