@@ -21,30 +21,30 @@ class OrganizationsControllerTest < ActionController::TestCase
 
   test 'should create unit' do
     assert_difference('Organization.count') do
-      post :create, organization: {
+      post :create, params: { organization: {
         code: 'NEW_UNIT', organization_id: @organization.id, name: SecureRandom.hex,
         organization_type: 'unit'
-      }
+      } }
     end
 
     assert_redirected_to organizations_path
   end
 
   test 'should show unit' do
-    get :show, id: @organization
+    get :show, params: { id: @organization }
     assert_response :success
   end
 
   test 'should get edit' do
-    get :edit, id: @organization
+    get :edit, params: { id: @organization }
     assert_response :success
   end
 
   test 'should update unit' do
-    patch :update, id: @organization, organization: {
+    patch :update, params: { id: @organization, organization: {
       code: @organization.code, organization_id: @organization.parent, name: @organization.name
-    }
-    assert_not flash.empty?
+    } }
+    assert !flash.empty?
     assert_redirected_to organizations_path
   end
 
@@ -53,7 +53,7 @@ class OrganizationsControllerTest < ActionController::TestCase
                            organization_id: @organization.id, organization_type: 'unit')
     org.save!
     assert_difference('Organization.count', -1) do
-      delete :destroy, id: org
+      delete :destroy, params: { id: org }
     end
 
     assert_redirected_to organizations_path
@@ -61,7 +61,7 @@ class OrganizationsControllerTest < ActionController::TestCase
 
   test 'should show error when cannot destroy unit with associated records' do
     assert_no_difference('Organization.count') do
-      delete :destroy, id: @organization
+      delete :destroy, params: { id: @organization }
     end
     assert_not flash.empty?
 
@@ -76,25 +76,25 @@ class OrganizationsControllerTest < ActionController::TestCase
       get :new
       assert_response :forbidden
 
-      get :show, id: @organization
+      get :show, params: { id: @organization }
       assert_response :forbidden
 
-      get :edit, id: @organization
+      get :edit, params: { id: @organization }
       assert_response :forbidden
 
-      post :create, unit: {
+      post :create, params: { unit: {
         code: 'NEW_UNIT', organization_id: @organization.organization_id, name: @organization.name,
         organization_type: 'unit'
-      }
+      } }
       assert_response :forbidden
 
-      patch :update, id: @organization, unit: {
+      patch :update, params: { id: @organization, unit: {
         code: @organization.code, organization_id: @organization.organization_id, name: @organization.name,
         organization_type: 'unit'
-      }
+      } }
       assert_response :forbidden
 
-      delete :destroy, id: @organization
+      delete :destroy, params: { id: @organization }
       assert_response :forbidden
     end
   end
